requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
SQLAlchemy==2.0.40
psycopg2-binary==2.9.10
gunicorn==23.0.0
beautifulsoup4==4.13.4
requests==2.32.3
python-dotenv==1.1.0
Flask-Session==0.8.0
